test(properites): pull setup into function transform unit into prop tests
setup being an ExUnit macro is only called once, so we pull its body
into an extra function which we can call when needed.

We've discovered that `async: true` does nothing for ExUnitProperties,
so one of the tests is commented out for now.

We've also split the use-case driven unit test into a better property
test shape